Split test suites into seperate files
